+++ akkad-springboot/springboot-dist-lock/dist-lock-redis/src/main/java/xyz/wongs/weathertop/deno/mapper/RedisLockMapper.java
@@ -6,9 +6,9 @@ import xyz.wongs.weathertop.deno.entity.RedisLock;
 public interface RedisLockMapper extends BaseMapper<RedisLock,Integer> {
     int deleteByPrimaryKey(Integer id);
 
-    int insert(RedisLock record);
+    Long insert(RedisLock record);
 
-    int insertSelective(RedisLock record);
+    Long insertSelective(RedisLock record);
 
     RedisLock selectByPrimaryKey(Integer id);
